[🚚change]: 캔버스 영역 컴포넌트 분리. 임시 더미데이터로 gameScore 프롭스전달
