+++ src/main.tsx
@@ -8,7 +8,7 @@ import AuthRoute from './common/AuthRoute/AuthRoute.tsx';
 import Header from './common/Header/Header.tsx';
 import BodyLayout from './common/Layout/BodyLayout.tsx';
 import Layout from './common/Layout/Layout.tsx';
-import CanvasTrack from './pages/GamePage/canvas/CanvasTrack.tsx';
+import TestIngamePage from './pages/GamePage/canvas/TestIngamePage.tsx';
 import GamePage from './pages/GamePage/GamePage.tsx';
 import KaKaoLoginPage from './pages/KaKaoLoginPage/KaKaoLoginPage.tsx';
 import LoginPage from './pages/LoginPage/LoginPage.tsx';
@@ -75,7 +75,7 @@ export const router = createBrowserRouter([
           },
           {
             path: '/test',
-            element: <CanvasTrack />,
+            element: <TestIngamePage />,
           },
           {
             path: '/*',
